+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/expression/common/ECallExpression.kt
@@ -29,7 +29,7 @@ class ECallExpression(
     override var type: Type,
     override var reference: Declaration,
     override var receiver: EExpression?,
-    override val isSafeAccess: Boolean,
+    override var isSafeAccess: Boolean,
     val valueArguments: ArrayList<EExpression>,
     var typeArguments: ArrayList<Type>
 ) : EReceiverExpression() {