+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/expression/common/EReceiverExpression.kt
@@ -24,7 +24,7 @@ abstract class EReceiverExpression : EExpression(), ExpressionContainer {
 
     abstract var reference: Declaration
     abstract var receiver: EExpression?
-    abstract val isSafeAccess: Boolean
+    abstract var isSafeAccess: Boolean
 
     override fun acceptChildren(visitor: TreeVisitor) {
         receiver?.accept(visitor)