+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/expression/common/EReferenceExpression.kt
@@ -18,6 +18,7 @@ package io.verik.compiler.ast.element.expression.common
 
 import io.verik.compiler.ast.common.Declaration
 import io.verik.compiler.ast.common.Type
+import io.verik.compiler.ast.element.declaration.common.EAbstractProperty
 import io.verik.compiler.ast.property.SerializationType
 import io.verik.compiler.common.Visitor
 import io.verik.compiler.message.SourceLocation
@@ -27,7 +28,7 @@ class EReferenceExpression(
     override var type: Type,
     override var reference: Declaration,
     override var receiver: EExpression?,
-    override val isSafeAccess: Boolean
+    override var isSafeAccess: Boolean
 ) : EReceiverExpression() {
 
     override val serializationType = SerializationType.EXPRESSION
@@ -38,5 +39,18 @@ class EReferenceExpression(
 
     override fun accept(visitor: Visitor) {
         visitor.visitReferenceExpression(this)
+    }
+
+    companion object {
+
+        fun of(property: EAbstractProperty): EReferenceExpression {
+            return EReferenceExpression(
+                property.location,
+                property.type.copy(),
+                property,
+                null,
+                false
+            )
+        }
     }
 }