+++ verik-compiler/src/main/kotlin/io/verik/compiler/transform/lower/ExpressionExtractorStage.kt
@@ -74,13 +74,7 @@ object ExpressionExtractorStage : ProjectStage() {
                 expression.location,
                 property
             )
-            val referenceExpression = EReferenceExpression(
-                expression.location,
-                property.type.copy(),
-                property,
-                null,
-                false
-            )
+            val referenceExpression = EReferenceExpression.of(property)
             return Pair(propertyStatement, referenceExpression)
         }
     }