+++ verik-compiler/src/main/kotlin/io/verik/compiler/transform/upper/ProceduralAssignmentTransformerStage.kt
@@ -72,13 +72,7 @@ object ProceduralAssignmentTransformerStage : ProjectStage() {
             val initializer = property.initializer
             return if (initializer != null) {
                 property.initializer = null
-                val referenceExpression = EReferenceExpression(
-                    property.location,
-                    property.type.copy(),
-                    property,
-                    null,
-                    false
-                )
+                val referenceExpression = EReferenceExpression.of(property)
                 val binaryExpression = EKtBinaryExpression(
                     property.location,
                     Core.Kt.C_Unit.toType(),
@@ -109,13 +103,7 @@ object ProceduralAssignmentTransformerStage : ProjectStage() {
             val initializer = property.initializer
             return if (initializer is ECallExpression && initializer.reference == Core.Vk.F_oni_Event_Event_Function) {
                 property.initializer = null
-                val referenceExpression = EReferenceExpression(
-                    property.location,
-                    property.type.copy(),
-                    property,
-                    null,
-                    false
-                )
+                val referenceExpression = EReferenceExpression.of(property)
                 val body = initializer.valueArguments.last().cast<EFunctionLiteralExpression>().body
                 if (body.isEmpty()) {
                     Messages.SEQ_ASSIGNMENT_NO_ONR_EXPRESSION.on(property)